Add OptimizeWMOs Class
Cleaned up development files to make a demo and class definitions for OptimizeWMOs and CostProblem

This is an MVP optimizer class.

This should be easy to build on for upcoming improvements.
